Move C code from comments to auxiliary file
